hcldec must use WithoutOptionalAttributesDeep
When hcldec needs to return a synthetic value, it must use
WithoutOptionalAttributesDeep to ensure the correct type for Null or
Unknown values. Optional attributes are normally removed from the value
type when decoding, but since hcldec is dealing with the decoder spec
directly, the optional attr types are still going to be present in these
cases.