hclparse: change format for path on error to open
When the hclparse's `Parser' attempts to open a file, and fails to do
so, we return a message that exposes the path with the `%s'
format-string specifier.

This specifier however duplicates the `\' in the path given, which on
Windows environments is common as it is the default delimiter between
directories.

As an alternative, Go offers the `%#q' specifier for format-strings,
which does not escape sequences as if they were go code, but instead
uses the alternative multiline syntax with '`', which leads to paths
not being over-escaped.