hclparse: add the error on failure to read file
When a file fails to be read from disk, we return an error, but we don't
include the real problem in the hcl.Diagnostics returned to the caller,
which can be problematic in order to understand why reading failed.
This commit fixes this problem by including the verbatim error message
in the Details of the returned diag.