Use Go 1.20
Versions before this are unsupported by the Go team anyway, but the main
reason is because our pull request checks workflow uses this to determine
which toolchain version to run the tests with.

This also brings in the new "go fmt" conventions for godoc comments, so
this commit updates them all to match those conventions.