Prevent diagnostic writes from panicking on marked values
Fixes https://github.com/hashicorp/hcl/issues/737

When the `DiagnosticTextWriter` writes out a diagnostic,
if the diagnostic has an Expression and EvalContext,
it will also write out the evaluation context
in which the diagnostic occurred.

The `cty.Value` of the variable relevant to the expression
will be rendered as a string by `valueStr`, but this does not
take into account marked values and will cause a panic by
`val.AsString()`, `val.LengthInt()`, etc.

This change prevents panics by always returning "(marked value)"
for marked values in `valueStr`. Whether this is always
appropriate is debatable, but it is better than
the current situation, which causes panics.